MemoryStore now uses fixed memory pool per store
MemoryStore now pre-allocates max_bytes from the eviction policy,
then installs a custom allocator in that pool that the store will
use. In testing, memory fragmentation appears to have dropped
significantly, significantly decreased the number of sys-calls
managing memory and slightly reduced user time.

In addition, we remove two full copies of our data when we store
it in the memory store.

closes: #289